Physics Labs
Analytical Mechanics: Modeled the motion of a spring pendulum in 2D.

Mechanical C: Used a laser and mirrors to calculate the speed of light.

Interferometer: Found the wavelength of a laser and the index of
refraction of air.

Thermoconductivity: Calculated the thermoconductivity of glass.

Xray: Used Bragg diffraction to find the wavelength of X-rays and the
density of MgO.